// types/admin-bot.ts
export interface AdminBotConfig {
  menuText: string;
  menuFooterText: string | null;
  panelButtonText: string;
  subscriptionButtonText: string;
  supportButtonText: string;
  menuImageUrl: string | null;
  menuImagePath: string | null;
  subscriptionHeaderText: string;
  subscriptionBodyText: string;
  subscriptionFooterText: string | null;
  subscriptionRenewButtonText: string;
  subscriptionChangeButtonText: string;
  subscriptionDetailsButtonText: string;
  subscriptionNoPlanHeaderText: string;
  subscriptionNoPlanBodyText: string;
  subscriptionNoPlanButtonText: string;
  subscriptionPlanListTitle: string;
  subscriptionPlanListBody: string;
  subscriptionPlanListButtonText: string;
  subscriptionPlanListFooterText: string | null;
}